requirements.txt
lancedb
langchain
pypdf
docx2txt
pyarrow
python-doctr[torch,viz,html,contrib]
torch
langchain-community
sentence-transformers
langgraph
langchain_openai
pandas
ipython
langchain-huggingface
requests
beautifulsoup4
newspaper3k
lxml[html_clean] 